Resolve Clang Compile warnings
Commit resolves select warnings that fill the build log with repetition when
using the flag -Weverything. Warnings highlighted areas where some improvements
to code could be made. Changes resulted in a greatly reduced log file size.

lib/errorloger.h -Wsign-conversion,Value Conversion Issue - two occurrences
FileLocation constructor overloads have in their signatures unsigned int for
line value. In context, line location is typed as int in the class. In context,
line is intended to be -1 to indicate no line but enforce 0 or > with the
unsigned value. Silence warning with static_cast<int> and refactor due to line
length.

lib/platform.h -Wsign-conversion,Value Conversion Issue
Repeated warnings occur when using functions in class(private and public)
functions. Simple task of silencing the repeative warnings with a static_cast
to appropriate type.

lib/platform.h -Wcovered-switch-default
'default' case in usage of PlatformType was flagged as possibly covering
conditions outside of defined PlatformType. Added check to handle out-of-bounds
values, removed default, but left switch untouched.

gui/resultstree.cpp -Wcovered-switch-default
Numerous warnings encountered with usage of default with Severity enum.
However, the code for the gui was improved by removing code that duplicates
the actions of a Severity class function.

+++ lib/errorlogger.h
@@ -193,12 +193,21 @@ public:
                 : fileIndex(0), line(0), col(0) {
             }
 
-            FileLocation(const std::string &file, unsigned int aline)
-                : fileIndex(0), line(aline), col(0), mOrigFileName(file), mFileName(file) {
+            FileLocation(const std::string &file, unsigned int aline):
+                fileIndex(0),
+                line(static_cast<int>(aline)),
+                col(0),
+                mOrigFileName(file),
+                mFileName(file) {
             }
 
-            FileLocation(const std::string &file, const std::string &info, unsigned int aline)
-                : fileIndex(0), line(aline), col(0), mOrigFileName(file), mFileName(file), mInfo(info) {
+            FileLocation(const std::string &file, const std::string &info, unsigned int aline):
+                fileIndex(0),
+                line(static_cast<int>(aline)),
+                col(0),
+                mOrigFileName(file),
+                mFileName(file),
+                mInfo(info) {
             }
 
             FileLocation(const Token* tok, const TokenList* tokenList);
